Move import helper to profile
